Remove unnecessary / incorrect `return` statement from javascript cmds
`->evaluateScript()` is officially evaluated as a top-level statement in
the browser, rather than within a function context, so a `return`
statement is technically incorrect (there is nothing to `return` from).

This causes an internal error in Chrome, although ChromeDriver detects
it and retries it as a self-executing inline function which works but
adds an extra script evaluation.

It appears that by contrast the WebDriver protocol *does* internally
require a return statement, but the SeleniumDriver adds one if missing
so we don't need to specify it ourselves.